moved more test over to the vm. Fixed len for []uint8
